ARM: head.S: Extend __fixup_smp to handle A9 MPCore with 1 core
The generic code is well equipped to differentiate between
SMP and UP configurations. However, there are some devices,
like AM437x from Texas Instruments, which use MPCore in a
single processor configuration. To enable these processors
to co-exist in a CONFIG_SMP=y build by leveraging the
SMP_ON_UP black magic we need to additionally check the
number the cores in MPCore configuration. Without such a
check in place, the startup code tries to execute ALT_SMP()
set of instructions which can lead to issues. One particular
example of this is the undefined instruction fault seen
on the pre-Si platform when the NS code tried to write to
ACTLR with NSACR.NS_SMP set to 0.

Since the additional check appears to be required only
on one platform, this patch extends the SMP detection
logic by adding in check for the no. of cores only for A9
MPCore systems by reading the SCU configuration register.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>
Signed-off-by: Sricharan R <r.sricharan@ti.com>