ARM: OMAP4: CM code: Remove the check for non-zero clkctrl_offs
AM335x which will migrate to the OMAP4 APIs in subsequent patches
has one CLKCTRL at offset 0. A similar check existed in the AM335x
custom APIs and that was removed to fix a crash on boot in commit
169c82a (ARM: OMAP2: am33xx-hwmod: Fix "register offset NULL check" bug).

Do the same in the OMAP4 API to avoid any regressions on AM335x
when it switches over to these APIs.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>