ARM: OMAP2+: AM437X: Basic PM support
Update the various files to add in support for AM437x.
I have kept the AM335x and AM437x under the appropriate
soc_is_*() checks but some things could be optimized further.

For now i have kept the mailbox based IPC mechaism for
both the SoCs - if the mailbox stuff is getting cleaned up
in mainline we might as well leverage that and cut down on
some lines of code.

There are some TODOs that have not gotten around to fixing
due the lack of time.

One thing that definitely needs to be reworked is the approach
to differentiate between the two SoCs in the assembly file.
The recent patches on cpuidle from Russ would be a good thing
to look at.

Oh and do make sure you pick up the latest M3 binary (0x24)
to test things on AM437x and AM335x.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>